% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phs_colour_values}
\alias{phs_colour_values}
\title{PHS colour names and their hex code.}
\format{
A character vector
}
\source{
\url{http://spark.publichealthscotland.org/media/2176/chart-and-dashboard-accessibility-guidance-version-12.pdf}
}
\usage{
phs_colour_values
}
\description{
A vector containing PHS colour names and their corresponding hex code.
It is used within \code{\link{phs_colours}}.
}
\details{
It contains PHS colour names in the PHS chart and dashboard guidance, such as PHS blue.
}
\keyword{datasets}
