% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_colours.R
\name{phs_colours}
\alias{phs_colours}
\alias{phs_colors}
\title{Return hex value for specific named colours}
\usage{
phs_colours(colourname = NULL, keep_names = FALSE)

phs_colors(colourname = NULL, keep_names = FALSE)
}
\arguments{
\item{colourname}{A colour name or vector of colour names.}

\item{keep_names}{whether to keep the colour names with hex values.
Defaulted as FALSE.}
}
\description{
\code{phs_colours} takes named PHS colours and returns their
hex value \code{\link{phs_colour_values}}
}
\details{
Named PHS colours can be phs-blue, phs-purple
}
\examples{
phs_colours("phs-blue")

phs_colours(c("phs-blue", "phs-purple"))

phs_colours()

phs_colors("phs-blue")

phs_colors(c("phs-blue", "phs-purple"))

phs_colors()
}
