% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_data.R
\docType{data}
\name{phs_colour_values}
\alias{phs_colour_values}
\alias{phs_palettes}
\alias{phs_palette_types}
\title{PHS colours and colour palettes}
\format{
\code{phs_colour_values} - A \code{character} vector of PHS colours and hex codes.

\code{phs_colour_palettes} - A \code{list} of PHS colours grouped into palettes.

\code{phs_palette_types} - A \code{list} of PHS palettes grouped into palette types.
}
\source{
\url{http://spark.publichealthscotland.org/media/2176/chart-and-dashboard-accessibility-guidance-version-12.pdf}
}
\usage{
phs_colour_values

phs_palettes

phs_palette_types
}
\description{
PHS colours and colour palettes
}
\details{
\subsection{PHS colours}{

PHS colour names and their corresponding hex codes.

\code{phs_colour_values} is a character vector that contains PHS colour names
in the PHS chart and dashboard guidance, such as PHS blue.
}

\subsection{PHS palettes}{

A list grouping PHS colours into palettes.
It is used within \code{\link{scale_colour_discrete_phs}},
\code{\link{scale_colour_continuous_phs}},
\code{\link{scale_fill_discrete_phs}} and
\code{\link{scale_fill_continuous_phs}}.

It contains PHS colour palettes such as main, main-purples.
}

\subsection{PHS palette types}{

A list grouping PHS palettes into PHS palette types.
It is used within \code{\link{scale_colour_discrete_phs}},
\code{\link{scale_colour_continuous_phs}},
\code{\link{scale_fill_discrete_phs}} and
\code{\link{scale_fill_continuous_phs}}.

It groups PHS palettes into two types - qualitative and sequential.

The palettes and their types are listed below.\tabular{ll}{
   \strong{Qualitative} \tab \strong{Sequential} \cr
   main \tab x \cr
   supporting \tab x \cr
   all \tab x \cr
   x \tab main-purples \cr
   x \tab main-magentas \cr
   x \tab main-blues \cr
   x \tab main-greens \cr
   x \tab supporting-graphites \cr
   x \tab supporting-teals \cr
   x \tab supporting-liberties \cr
   x \tab supporting-rusts \cr
}

}
}
\keyword{datasets}
