% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phs_palette_types}
\alias{phs_palette_types}
\title{PHS palette types}
\format{
A character list
}
\usage{
phs_palette_types
}
\description{
A list grouping PHS palettes into PHS palettes types.
It is used within \code{\link{scale_colour_discrete_phs}}.
}
\details{
It groups PHS palettes into two types - qualitative and sequential.
}
\keyword{datasets}
