% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_theme.R
\name{phs_theme}
\alias{phs_theme}
\title{Public Health Scotland theme for ggplot2 plots.}
\usage{
phs_theme()
}
\description{
This adds certain styles to ggplot2 plots that conform with
Public Health Scotland style guidelines.
}
\examples{
ggplot2::qplot(mpg, wt, data = mtcars) + phs_theme()
}
