% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_colours.R
\name{phs_colours}
\alias{phs_colours}
\title{Return hex value for specific named colours}
\usage{
phs_colours(colourname = NULL)
}
\arguments{
\item{colourname}{A colour name or vector of colour names.}
}
\description{
\code{phs_colours} takes named PHS colours and returns their
hex value \code{\link{phs_colour_values}}
}
\details{
Named PHS colours can be phs-blue, phs-darkblue
}
\examples{
phs_colours("phs-blue")

phs_colours(c("phs-blue", "phs-darkblue"))

phs_colours()

}
