% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phs_palette_types}
\alias{phs_palette_types}
\title{PHS palette types}
\format{
A character list
}
\usage{
phs_palette_types
}
\description{
A list grouping PHS palettes into PHS palettes types.
It is used within \code{\link{scale_colour_discrete_phs}},
\code{\link{scale_colour_continuous_phs}},
\code{\link{scale_fill_discrete_phs}} and
\code{\link{scale_fill_continuous_phs}}.
}
\details{
It groups PHS palettes into two types - qualitative and sequential.
The palettes and their types are listed below.\tabular{ll}{
   \strong{Qualitative} \tab \strong{Sequential} \cr
   main \tab x \cr
   supporting \tab x \cr
   all \tab x \cr
   x \tab main-purples \cr
   x \tab main-magentas \cr
   x \tab main-blues \cr
   x \tab main-greens \cr
   x \tab supporting-graphites \cr
   x \tab supporting-teals \cr
   x \tab supporting-liberties \cr
   x \tab supporting-rusts \cr
}
}
\keyword{datasets}
