% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{phs_palettes}
\alias{phs_palettes}
\title{PHS palettes}
\format{
A character list
}
\usage{
phs_palettes
}
\description{
A list grouping PHS colours into palettes.
It is used within \code{\link{scale_colour_discrete_phs}}.
}
\details{
It contains PHS colour palettes such as main, main-purples.
}
\keyword{datasets}
