% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phsstyles.R
\docType{package}
\name{phsstyles}
\alias{phsstyles}
\title{\code{phsstyles} package}
\description{
Standard graphic styles for use in PHS.
}
\details{
See the README on
\href{https://github.com/Public-Health-Scotland/phsstyles#readme}{GitHub}.
}
